% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{actiData}
\alias{actiData}
\alias{is_actiData}
\title{Storage of estimates of regulation of activity of metabolic reactions: \code{actiData} class.}
\usage{
actiData(reg, mc = NULL, ...)

is_actiData(x)
}
\arguments{
\item{reg}{a data frame with estimates of regulation of metabolic reactions
as described in Details.}

\item{mc}{\code{NULL} or a numeric matrix with estimates of activity of metabolic
reactions obtained in Monte-Carlo simulations.}

\item{...}{additional arguments, currently none.}

\item{x}{an object}
}
\value{
\code{actiData()}: a list of class \code{actiData} as described in Details,
\code{is_actiData()}: a logical value indicating if an object is an instance of
\code{actiData} class.
}
\description{
A list for storage of estimates of regulation of activity of metabolic reactions
and, optionally, results of Monte Carlo simulations of reaction activity.
}
\details{
\code{actiData} class objects are lists with the obligatory element \code{reg},
a data frame which stores the reaction regulation estimates, and \code{mc},
a numeric matrix of regulation estimates in single iterations of Monte Carlo
simulations.
The \code{reg} data frame is expected to have the following obligatory columns:
\itemize{
\item \strong{id} and \strong{name}: character variables storing identifiers of metabolic reactions
\item \strong{subsystem}: a character variable with names of Recon subsystems
\item \strong{fold_reg}: a numeric variable with estimates of fold-regulation of reaction
activity.
}

The optional columns in \code{reg} are:
\itemize{
\item \strong{error}, \strong{lower_ci}, and \strong{upper_ci}: numeric variables with errors,
lower and upper bounds of confidence intervals of fold-regulation of reaction activity.
\item \strong{z}: a numeric variable with values of the test statistic Z
\item \strong{p_value} and \strong{p_adjusted}:  numeric variables with raw p values and
p-values adjusted for multiple testing.
}
}
