% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{annotate_reactions}
\alias{annotate_reactions}
\title{Annotate reaction identifiers with information from a \code{reactDB} database.}
\usage{
annotate_reactions(x, database)
}
\arguments{
\item{x}{a data frame with column \code{id} with identifiers of reactions.}

\item{database}{an object of class \code{\link{reactDB}} created with
functions \code{\link{extract_genes}} or \code{\link{as_reactDB}}.}
}
\value{
the input data frame appended with columns \code{name} and \code{subsystem} with
reaction names and subsystem assignment.
}
\description{
Appends a data frame with identifiers of metabolic reactions with
reaction names and Recon subsystem information.
Intended for internal use only.
}
