% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_funs.R
\name{build_geneSBML}
\alias{build_geneSBML}
\title{Create a SBML model from gene Entrez IDs (not active anymore).}
\usage{
build_geneSBML(x, ...)
}
\arguments{
\item{x}{a vector with gene regulation estimates named
with Entrez ID identifiers.}

\item{...}{additional argument, currently none.}
}
\description{
Function \code{bulid_geneSBML()} is not active anymore and was replaced by
much more robust \code{\link{get_regulation}} with a slightly different
interface and output.
Calling \code{bulid_geneSBML()} returns \code{NULL} with a warning.
}
