% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{calculate.geneSBML}
\alias{calculate.geneSBML}
\title{Reaction regulation rates.}
\usage{
\method{calculate}{geneSBML}(x, react_id = x$reg$react_id, ...)
}
\arguments{
\item{x}{a geneSBML object.}

\item{react_id}{a character vector of reaction IDs.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with reaction regulation estimates and, optionally,
their errors.
}
\description{
Extracts regulation rates and their errors from and geneSBML object.
}
