% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_funs.R
\name{check_geneSBML}
\alias{check_geneSBML}
\title{Check if reactions or metabolites are present in the geneSBML model -------}
\usage{
check_geneSBML(react_id = NULL, metab_id = NULL, geneSBML)
}
\arguments{
\item{react_id}{BiGG reaction ID, with or without the leading 'R_' string.}

\item{metab_id}{BiGG metabolite ID, with or without the leading 'M_' string.}

\item{geneSBML}{a geneSBML object.}
}
\value{
a list of logical vectors logical vectors, each for the reactions
and metabolite IDs, if both the reactions and metabolites are queried.
Otherwise a named vector with logical values.
}
\description{
Checks if the given reaction can be found in the geneSBML model.
}
