% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiData_methods.R
\name{components.actiData}
\alias{components.actiData}
\title{Access components of \code{actiData} objects.}
\usage{
\method{components}{actiData}(object, type = c("reg", "mc"), ...)
}
\arguments{
\item{object}{a \code{\link{actiData}} object.}

\item{type}{type of the output, see above.}

\item{...}{extra arguments, currently none.}
}
\value{
the output specified by \code{type} argument:
a data frame with reaction identifiers, names, subsystems and regulation
estimates (\code{type = "reg"}), a numeric matrix with reaction activity estimates
in Monte Carlo simulations (\code{type = "mc"}).
}
\description{
Extracts the \code{reg} and \code{mc} components of the object.
}
