% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{components.geneSBML}
\alias{components.geneSBML}
\title{Extract geneSBML model components.}
\usage{
\method{components}{geneSBML}(
  object,
  type = c("model", "regulation", "gene_map", "reactions", "metabolites", "genes"),
  ...
)
}
\arguments{
\item{object}{a geneSBML object.}

\item{type}{type of the output, see above.}

\item{...}{extra arguments, currently none.}
}
\value{
the output specified by 'type' argument.
}
\description{
Extracts the following geneSBML components:
SBML model (type = 'model', default),
reaction regulation estimates ('regulation'),
gene mapping ('gene_map'),
all reaction IDs ('reactions'),
all metabolite IDs ('metabolites')
and all gene IDs ('genes').
}
