% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{count.geneSBML}
\alias{count.geneSBML}
\title{Count regulated reactions per subsystem.}
\usage{
\method{count}{geneSBML}(x, signif_type = c("fdr", "raw"), regulation_level = 1, ...)
}
\arguments{
\item{x}{an object of the geneSBML class.}

\item{signif_type}{type of significance to be used for definition of
regulated reactions:
'raw' or 'fdr' (FDR-corrected, default).
Ignored if the model contains no error information.}

\item{regulation_level}{fold-regulation level cutoff used for definition
of regulated reactions.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with reaction counts according
to their regulation status.
}
\description{
Computes counts of significantly activated, inhibited and
all reactions per subsystem.
}
