% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiData_methods.R
\name{count_regulated}
\alias{count_regulated}
\alias{count_regulated.actiData}
\title{Count regulated reactions per subsystem.}
\usage{
count_regulated(x, ...)

\method{count_regulated}{actiData}(x, ...)
}
\arguments{
\item{x}{a \code{\link{actiData}} object.}

\item{...}{additional arguments passed to methods.}
}
\value{
a data frame with reaction counts according
to their regulation status.
}
\description{
Computes counts of significantly activated and inhibited reactions:
in the whole database and for particular subsystems.
Please make sure that the object contains activity regulation status
information.
If not, please call \code{\link{identify_regulated}} prior to
counting the significant effects.
}
\details{
If the object has no regulation status information (column \code{regulation} in the
\code{reg} element), an error is raised.
}
\seealso{
\code{\link[=get_regulation]{get_regulation()}}
}
