% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{draw_norm}
\alias{draw_norm}
\title{Draw values from the gene regulation estimate distribution.}
\usage{
draw_norm(x, err, n, scale = c("identity", "log2", "log"), seed = NULL)
}
\arguments{
\item{x}{a numeric vector with fold-regulation estimates.
Elements are named with Entrez IDs.}

\item{err}{a numeric vector with an error estimate such as SD or SEM,
named with Entrez IDs.}

\item{n}{number of values to be drawn.}

\item{scale}{name of the scale of the provided regulation estimates and
their errors: "identity" (default), "log2", or "log".}

\item{seed}{seed of the random number generator.}
}
\value{
a numeric matrix: genes are columns, their regulation estimates are
}
\description{
Draws values from normal distributions of gene regulation estimates.
The gene expression regulation estimate serves as the mean of the distribution,
the gene expression regulation error serves as the standard deviation of
the distribution.
}
