% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{escape_numbers}
\alias{escape_numbers}
\title{Escape Entrez ID numbers in a string.}
\usage{
escape_numbers(str)
}
\arguments{
\item{str}{a string.}
}
\value{
a sting.
}
\description{
Finds Entrez ID in a string and escapes them with '``.
}
