% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_rule}
\alias{eval_rule}
\title{Evaluate gene mapping rules.}
\usage{
eval_rule(x, rule_exp, parent_env)
}
\arguments{
\item{x}{a vector with gene regulation estimates named
with Entrez ID identifiers.}

\item{rule_exp}{a list of gene mapping rules.}

\item{parent_env}{an environments containing evaluation functions and values
for genes missing from the x gene regulation estimate vector.}
}
\value{
a numeric vector with the reaction regulation estimates.
Named with reaction IDs.
}
\description{
Evaluates a list of gene mapping rules stored as expression.
Such rules are retrieved from the SBML database e.g.
with \code{\link{extract_genes}}.
}
