% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eval_rules}
\alias{eval_rules}
\title{Evaluate gene mapping rules.}
\usage{
eval_rules(x, database, parent_env)
}
\arguments{
\item{x}{a vector with gene regulation estimates named
with Entrez ID identifiers.}

\item{database}{an object of class \code{\link{reactDB}} created with
functions \code{\link{extract_genes}} or \code{\link{as_reactDB}}.}

\item{parent_env}{an environments containing evaluation functions and values
for genes missing from the x gene regulation estimate vector.}
}
\value{
a numeric vector with the reaction regulation estimates.
Named with reaction IDs.
}
\description{
Evaluates a list of gene mapping rules stored as expression.
Such rules are retrieved from data frames of reaction annotations
with \code{\link{extract_genes}} or \code{\link{as_reactDB}}.
}
