% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_funs.R
\name{extract_genes}
\alias{extract_genes}
\alias{as_reactDB}
\title{Extract gene identifiers from reaction annotations.}
\usage{
extract_genes(x, react_id = NULL, inspect_errors = FALSE)

as_reactDB(x, ...)
}
\arguments{
\item{x}{a data frame with the following obligatory columns: \code{id} with reaction
identifiers, \code{name} with reaction names, \code{subsystem} with subsystem assignment,
and \code{gene_association} with character strings with gene association rules.}

\item{react_id}{BiGG reaction ID, with or without the leading 'R_' string.
Defaults to \code{NULL}, which means that all reactions are mapped to genes.}

\item{inspect_errors}{logical. If \code{TRUE}, the function returns parsing errors.}

\item{...}{additional arguments passed to \code{extract_genes()}.}
}
\value{
a data frame of class \code{\link{reactDB}} containing reaction
IDs (\code{id}), reaction names (\code{name}), subsystem information (\code{subsystem}),
list of gene identifiers (\code{entrez_id}) and a list of expressions to be
evaluated by at calculating reaction activity estimates.
}
\description{
The functions extract Entrez ID gene identifiers mapped to all or
user-specified reactions present in a data frame with reaction annotation
and gene - reaction association rules.
Entrez IDs of genes associated with reactions are listed, and character
strings with the gene - reaction association rules are translated to R
expressions.
\code{as_reactDB()} offers a simplified tools for generation of
\code{\link{reactDB}} data frames,
while \code{extract_genes()} allows for selection of reactions of interest and
parsing error diagnostics.
}
\details{
Association rules with unrecognized Entrez IDs are removed from the output
and a parsing warning is raised.
The gene association rules have to operate with "bare" Entrez IDs without
the version information (i.e. numbers after a dot).
}
\references{
King ZA, Lu J, Dräger A, Miller P, Federowicz S, Lerman JA, Ebrahim A,
Palsson BO, Lewis NE. BiGG Models: A platform for integrating,
standardizing and sharing genome-scale models.
Nucleic Acids Res (2016) 44:D515–D522. doi:10.1093/NAR/GKV1049
}
