% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_funs.R
\name{extract_subsystems}
\alias{extract_subsystems}
\title{Extract subsystem assignment from a database.}
\usage{
extract_subsystems(database, react_id = NULL, as_list = FALSE)
}
\arguments{
\item{database}{an object of class SBMLDocument.}

\item{react_id}{BiGG reaction ID, with or without the leading 'R_' string.
Defaults to NULL, which means that all reactions are mapped.}

\item{as_list}{logical, should a list be returned? Defaults to FALSE.}
}
\value{
if 'as_list' is set to FALSE a data frame with the reaction ID
('react_id') and subsystem variable ('subsystem').
Otherwise a subsystem-named list of reaction ID vectors.
}
\description{
Extracts mapping of all or user-specified reactions to
subsystems defines in the given database SBML database.
}
