% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{fit.geneSBML}
\alias{fit.geneSBML}
\title{Fit a LIM based on geneSBML gene regulation data.}
\usage{
\method{fit}{geneSBML}(
  object,
  maximize,
  inequalities,
  constraints,
  externals,
  file.name = "model.lim",
  ...
)
}
\arguments{
\item{object}{a geneSBML object.}

\item{maximize}{a reaction or a series thereof to be maximized provided as
a single string or a character vector. Each reaction specified has
to be present in the regulation data stored in the geneSBML model.}

\item{inequalities}{a list specifying inequality constraints on the system,
see: \code{\link[BiGGR]{createLIMFromSBML}}.}

\item{constraints}{model constraints.}

\item{externals}{a character vector of metabolites for which the flux
balance analysis (FBA) has to be performed.}

\item{file.name}{a path and name of the LIM file.}

\item{...}{extra arguments, currently none.}
}
\value{
A model file with with extension ".lim" is created.
}
\description{
Fits a linear inverse model (LIM) to the SBML metabolic
reaction network model and reaction regulation estimates stored in the
geneSBML object.
Technically, a wrapper around \code{\link[BiGGR]{createLIMFromSBML}} with an
additional entry check. See the documentation of the genuine BiGGR function
for details. Reaction regulation estimates are fed into the LIM as
'equations' argument.
}
