% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_funs.R
\name{gene_to_react}
\alias{gene_to_react}
\title{Map gene EntrezIDs to reactions.}
\usage{
gene_to_react(entrez_id, geneSBML, lead_str = TRUE)
}
\arguments{
\item{entrez_id}{a vector of of Entrez IDs.}

\item{geneSBML}{a geneSBML object.}

\item{lead_str}{logical, should the reaction ID contain the leading 'R_'
string? Defaults to TRUE.}
}
\value{
a data frame with Entrez IDs and a list of the associated
reaction BiGG IDs.
}
\description{
Maps gene Entrez IDs to the associated reactions.
}
