% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_funs.R
\name{get_activity}
\alias{get_activity}
\alias{get_activity.matrix}
\alias{get_activity.data.frame}
\title{Calculate reaction activity scores based on gene expression data.}
\usage{
get_activity(x, ...)

\method{get_activity}{matrix}(
  x,
  database,
  or_fun = c("mean", "median", "min", "max"),
  and_fun = c("min", "max", "mean", "median"),
  x_default = 1,
  as_data_frame = TRUE,
  .parallel = FALSE,
  ...
)

\method{get_activity}{data.frame}(
  x,
  database,
  id_col = NULL,
  variables = NULL,
  or_fun = c("mean", "median", "min", "max"),
  and_fun = c("min", "max", "mean", "median"),
  x_default = 1,
  as_data_frame = TRUE,
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{x}{a numeric matrix or a data frame with gene expression metrics; genes
in columns, samples in rows.}

\item{...}{additional arguments, currently none.}

\item{database}{an object of class \code{\link{reactDB}} created with
functions \code{\link{extract_genes}} or \code{\link{as_reactDB}}.}

\item{or_fun}{one of "mean" (default), "median", "min", or "max".
Name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{and_fun}{one of "mean", "median", "min" (default), or "max".
Name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{x_default}{the default numeric value or \code{NA} for regulation of the genes
present in the database and absent from the regulation data. If \code{NA},
the genes missing from \code{x} won't be used in calculation of the reaction
activity estimates.}

\item{as_data_frame}{logical, should the matrix of reaction activity scores
be coerced to a data frame?}

\item{.parallel}{logical, should th computation be run in parallel?}

\item{id_col}{\code{NULL} or a character string specifying a variable serving as
a sample identifier. If \code{id_col = NULL}, the sample identifiers will be
extracted from row names of \code{x}.}

\item{variables}{\code{NULL} or a character vector specifying gene expression
variables used for calculation of reaction activity scores. If \code{NULL}, all
variables in \code{x} except of the sample identifier will be used.}
}
\value{
a numeric matrix or a data frame (if \code{as_data_frame = TRUE}) with
reaction activity scores.
Metabolic reactions named with their identifiers are in columns,
samples are in rows.
In the data frame output, sample identifiers are stored in the first column
named \code{sample_id}.
}
\description{
Function \code{get_activity()} calculates activity scores of metabolic reactions
by evaluation of gene - reaction association rules in a reaction annotation
database given gene expression data.
}
\details{
Function \code{get_activity()} takes a numeric matrix or a numeric data frame \code{x}
with gene expression metrics to calculate activity scores of metabolic reactions.
The scores are calculated by simple evaluation of gene - reaction association
rules stored in a reaction annotation database data frame of class
\code{\link{reactDB}} (created with \code{\link{as_reactDB}} or
\code{\link{extract_genes}}).

Of note, the gene expression metrics in \code{x} are not pre-processed prior to
calculation of the reaction activity scores; the pre-processing steps are
intended to be done by the user.
Because evaluation of the gene - reaction association rules uses simple
arithmetic with means, medians, minima or maxima of gene expression values,
it's highly recommended to bring the expression values in \code{x} approximate
equal scales, e.g. by computing Z-scores with \code{\link[microViz]{zScores}}.
}
\references{
King ZA, Lu J, Dräger A, Miller P, Federowicz S, Lerman JA, Ebrahim A,
Palsson BO, Lewis NE. BiGG Models: A platform for integrating,
standardizing and sharing genome-scale models.
Nucleic Acids Res (2016) 44:D515–D522. doi:10.1093/NAR/GKV1049
}
\seealso{
\code{\link[=extract_genes]{extract_genes()}}, \link{as_reactDB}
}
