% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_funs.R
\name{get_regulation}
\alias{get_regulation}
\title{Calculate reaction activity estimates based on differential gene expression.}
\usage{
get_regulation(
  x,
  err = NULL,
  database,
  scale = c("identity", "log2", "log"),
  or_fun = c("mean", "median", "min", "max"),
  and_fun = c("min", "max", "mean", "median"),
  x_default = 1,
  err_method = c("norm", "mc"),
  return_mc = FALSE,
  n_iter = 1000,
  mc_estimate = c("mean", "median"),
  burn_in = 0,
  ci_method = c("bca", "perc"),
  seed = NULL,
  .parallel = TRUE
)
}
\arguments{
\item{x}{a vector with gene regulation estimates named
with Entrez ID identifiers.}

\item{database}{an object of class \code{\link{reactDB}} created with
functions \code{\link{extract_genes}} or \code{\link{as_reactDB}}.}

\item{scale}{specifies the scale of the provided expression regulation and
error estimates: "identity" (default), "log2", or "log".}

\item{or_fun}{one of "mean" (default), "median", "min", or "max".
Name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{and_fun}{one of "mean", "median", "min" (default), or "max".
Name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{x_default}{the default numeric value or \code{NA} for regulation of the genes
present in the database and absent from the regulation data. If \code{NA},
the genes missing from \code{x} won't be used in calculation of the reaction
activity estimates.}

\item{err_method}{specifies the method of calculation of errors for reactions
as described in Details: "norm" (from normal distribution, default) or
"mc" (Monte Carlo simulation).}

\item{return_mc}{logical, should all Monte Carlo-determined regulation
estimates for reactions be returned? Defaults to \code{FALSE}.}

\item{n_iter}{number of iterations for the Monte Carlo simulation of
the reaction errors.}

\item{mc_estimate}{statistic used to compute the reaction activity estimate
in Monte Carlo simulation: either mean or median over algorithm iterations.
Ignored if no errors were provided or \code{err_method = "norm"}.}

\item{burn_in}{number of Monte Carlo iterations to be discarded prior to
computation of activity estimates and activity regulation errors. Ignored
if no errors were provided or \code{err_method = "norm"}.}

\item{ci_method}{method of calculation of confidence intervals:
BCA ("bca", default) or percentile ("perc").}

\item{seed}{a seed for the random number generator.}

\item{.parallel}{logical, should th computation be run in parallel?}
}
\value{
The function returns a list with the following elements:
\itemize{
\item \strong{reg} : a data frame with reaction identifiers (\code{id}),
fold-regulation estimates of reaction activity (\code{fold_reg}), estimates of
error and 95\% confidence intervals of reaction activity (only if \code{err} vector was
provided, columns: \code{error}, \code{lower_ci}, \code{upper_ci}), testing statistic Z (\code{z}, only
if \code{err} was provided and \code{err_method = "norm"}), raw p values and p values
adjusted for multiple testing with the Benjamini-Hochberg method
(only if \code{err} vector was provided, columns: \code{p_value} and \code{p_adjusted}).
\item \strong{mc}: returned only if \code{err_method = "mc"} and \code{return_mc = TRUE}, a numeric
matrix with estimates of fold-regulation of reaction activity in iterations of
the simulation algorithm. Reactions are in columns, iterations are in rows.
}
}
\description{
Intended for internal use.
Function \code{get_regulation()} computes fold-regulation estimates of metabolic
reaction activity given estimates of differential gene-expression with their
errors and gene - reaction association rules in a metabolic reaction annotation
database object..
If the errors are provided, statistical significance and confidence intervals
of the reaction activity are computed with two methods: (1) derived from normal
distribution of the error sums or (2) computed by serial evaluation of the
gene - reaction association rules for random draws from normal distributions
of the estimates of differential gene expression.
In the later case, the differential gene expression estimate serves as the
distribution mean and the error of differential gene expression serves as
the standard deviation.
}
