% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{Human_GEM_2_0_0}
\alias{Human_GEM_2_0_0}
\title{Reaction association rules for the Human-GEM 2.0.0 model.}
\format{
a data frame with 12931 rows and 8 columns:
\itemize{
\item \strong{id} Metabolic Atlas reaction identifiers
\item \strong{name} reaction name
\item \strong{ec_number} EC identifier of enzymatic reactions
\item \strong{gene_association} gene - reaction association rules as character strings
\item \strong{subsystem} Metabolic Atlas subsystem
\item \strong{confidence_level} confidence level for the gene - reaction association rule
\item \strong{reference} literature references
\item \strong{miriam} cross-references to other databases
}
}
\source{
\href{https://github.com/SysBioChalmers/Human-GEM}{GitHub repository of Metabolic Atlas} and
\href{https://github.com/PiotrTymoszuk/RECON_processing}{RECON_processing GitHub repository}
}
\usage{
data(Human_GEM_2_0_0)
}
\description{
Cleared and harmonized reaction annotation data for
\href{https://github.com/SysBioChalmers/Human-GEM}{Human-GEM 2.0.0 model of metabolism}.
The data were processed with the \href{https://github.com/PiotrTymoszuk/RECON_processing}{RECON_processing}
R pipeline.
}
\references{
Robinson JL, Kocabaş P, Wang H, Cholley PE, Cook D, Nilsson A, Anton M,
Ferreira R, Domenzain I, Billa V, et al.
An atlas of human metabolism. Sci Signal (2020) 13:eaaz1482–eaaz1482.
doi:10.1126/SCISIGNAL.AAZ1482
}
