% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiData_methods.R
\name{identify_regulated}
\alias{identify_regulated}
\alias{identify_regulated.actiData}
\title{Identity significantly regulated reactions.}
\usage{
identify_regulated(x, ...)

\method{identify_regulated}{actiData}(
  x,
  p_type = c("adjusted", "raw"),
  p_cutoff = 0.05,
  fold_cutoff = 1,
  ...
)
}
\arguments{
\item{x}{a \code{\link{actiData}} object.}

\item{...}{extra arguments passed to methods.}

\item{p_type}{type of p-values to be used for definition of
regulated reactions: "raw" or "adjusted" (FDR-corrected, default).}

\item{p_cutoff}{cutoff of the p value used for definition of regulated
reactions, see Details.}

\item{fold_cutoff}{fold-regulation level cutoff used for definition
of regulated reactions.}
}
\value{
a \code{\link{actiData}} object.
The \code{reg} element of the object is appended with additional columns:
\code{p_cutoff} (cutoff of the p-value), \code{p_type} (type of the p value used to
identify significant effects), \code{fold_cutoff} (cutoff of the fold-regulation estimate),
\code{regulation} (regulation status of the reaction: activated, inhibited, or
non-significant/ns).
}
\description{
Identifies significantly activated and inhibited reactions in
a \code{\link{actiData}} objects based on a cutoff of raw/adjusted p value
and activity estimate.
}
\details{
This method works only with \code{\link{actiData}} which store statistical
significance information (i.e. columns \code{p_value} and \code{p_adjusted}).
If this information is absent a warning is raised and regulation status
is inferred from the fold-regulation estimate only.

The \code{fold_cutoff} arguments defines the fold-regulation cutoff used for
identification of significantly regulated reactions.
Reactions with regulation estimates lower than \code{1/fold_cutoff} and passing
the significance criterion are identified as inhibited.
Reactions with regulation estimates higher than \code{fold_cutoff} and passing
the significance criterion are identified as inhibited.
}
\seealso{
\code{\link[=get_regulation]{get_regulation()}}
}
