% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{is_geneSBML}
\alias{is_geneSBML}
\alias{is_SBML}
\title{Test for the geneSBML and SBML class instance.}
\usage{
is_geneSBML(x)

is_SBML(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
The function test whether the object is an instance of the geneSBML class
or SBML class.
}
