% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{metabolites}
\alias{metabolites}
\title{BiGG metabolite annotation data.}
\format{
A data frame with 15724 rows and 6 variables:
\itemize{
\item \strong{bigg_id}: BiGG metabolite identifier
\item \strong{universal_bigg_id}: universal BiGG metabolite identifier
\item \strong{name}: metabolite name
\item \strong{model_list}: list of models with containing the metabolite
\item \strong{database_links}: links to the BiGG on-line database
\item \strong{old_bigg_ids}: legacy BiGG ID
}
}
\source{
BiGG: http://bigg.ucsd.edu/data_access
}
\usage{
data(metabolites)
}
\description{
A data frame with BiGG database annotation of metabolites.
}
