% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiData_methods.R
\name{plot.actiData}
\alias{plot.actiData}
\title{Diagnostic plots and visualization of reaction activity in \code{actiData} objects.}
\usage{
\method{plot}{actiData}(x, plot_type = c("errors", "numbers", "mc"), ...)
}
\arguments{
\item{x}{an \code{\link{actiData}} object.}

\item{plot_type}{plot type, see Details.}

\item{...}{additional arguments passed to the plotting functions specified
in Details.}
}
\value{
a \code{ggplot} graphic object.
}
\description{
Diagnostic plots and visualization of estimates of regulation of reaction
activity - the plotting method for \code{\link{actiData}} objects.
}
\details{
The \code{plot()} method for \code{\link{actiData}} objects generates the following
types of plots specified by \code{plot_type} argument:
\itemize{
\item \code{plot_type = "errors"}: scatter plot of regulation estimates and their errors
made with \code{\link{plot_errors}}.
\item \code{plot_type = "numbers"}: bar/stack plots of numbers or percentages of
differentially regulated
reactions in subsystems generated with \code{\link{plot_numbers}}. Please note
that the regulation status of metabolic reactions in \code{x} has to be determined
prior to plotting by calling \code{\link{identify_regulated}}.
\item \code{plot_type == "mc"}: plots of reactions regulation estimates in single
iterations of the Monte Carlo simulation with \code{\link{plot_mc}}.
}
}
\seealso{
\code{\link[=get_regulation]{get_regulation()}}
}
