% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{plot.geneSBML}
\alias{plot.geneSBML}
\title{Plot features of a geneSBML object.}
\usage{
\method{plot}{geneSBML}(
  x,
  type = c("regulation", "errors", "top", "volcano", "forest", "top_forest", "mc"),
  relevant.reactions = components(x, "reactions"),
  n_top = 10,
  show_fit = TRUE,
  point_size = 2,
  point_color = "black",
  point_alpha = 0.25,
  line_alpha = 1,
  signif_type = c("raw", "fdr"),
  regulation_level = 1,
  label_names = FALSE,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{x}{an object of the geneSBML class.}

\item{type}{type of the plot: 'regulation' (default), 'errors', 'top',
'volcano', 'top_forest' and 'forest', as specified above.}

\item{relevant.reactions}{a vector of reaction IDs to be presented
in the plots, defaults to all available reactions.}

\item{n_top}{number of top up- and downregulated reactions to be plotted
(type 'top' and 'forest') or top significant reactions to be labeled in
the Volcano plot.}

\item{show_fit}{logical, should a trend line be plotted? valid only for
the error plots and ignored otherwise.}

\item{point_size}{size of the points in the errors, Volcano or Forest plots.
Ignored for other plot types'.}

\item{point_color}{color of the points in the error plots.
Ignored for other plot types.}

\item{point_alpha}{alpha of the points in the error and volcano plots.
Ignored for other plot types.}

\item{line_alpha}{alpha of the lines connecting estimates in the MC plots.
Ignored for other plot types.}

\item{signif_type}{significance to be plotted in Volcano or Forest plots:
'raw' (default) or 'fdr' (FDR-corrected).}

\item{regulation_level}{fold-regulation level cutoff to be presented in
Forest and Volcano plots, defaults to 1.}

\item{label_names}{logical, should the reactions be labeled with their
full names retrieved from the model? Defaults to FALSE.}

\item{cust_theme}{custom ggplot2 theme.}

\item{...}{extra arguments: for 'errors' arguments controlling the
appearance of the trend line passed to \code{\link[ggplot2]{geom_smooth}};
for Volcano: arguments passed to \code{\link[microViz]{plot_volcano}};
for Forest plots: arguments passed to  \code{\link[microViz]{plot_top}} or
\code{\link[microViz]{plot_forest}}.}
}
\value{
a ggplot2 graphic object.
}
\description{
Plots features of the geneSBML object as specified by the 'type'
argument. For 'type' set to 'regulation' a bar plot of the
regulation values is created, 'errors' returns a point plot of regulation
estimates and regulation errors if errors can be retrieved from the model.
For 'type' set to 'top' a bar plot with top up- and downregulated reactions
is returned.
For 'volcano' a Volcano plot of raw or FDR-corrected p values and reaction
regulation estimates is returned.
The 'top_forest' draws a Forest plot of regulation estimates
along with 95\% confidence intervals for the top activated and
inhibited reactions.
The 'forest' option draws a Forest plot of regulation estimates
along with 95\% confidence intervals.
Finally, the type argument set to 'mc' draws Monte Carlo estimates of
pathway regulation in subsequent runs; this options is available only for
geneSBML model with errors estimated by the Monte Carlo method.
}
