% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{plot_errors}
\alias{plot_errors}
\title{Scatter plot of errors and regulation estimates for metabolic reactions.}
\usage{
plot_errors(
  x,
  reactions = NULL,
  subsystems = NULL,
  fun = identity,
  point_color = "steelblue",
  point_size = 2,
  point_alpha = 0.75,
  show_trend = TRUE,
  line_color = "black",
  line_alpha = 1,
  show_vline = TRUE,
  vline_color = "black",
  vline_type = "dashed",
  vline_alpha = 1,
  show_hline = TRUE,
  hline_color = "black",
  hline_type = "dashed",
  hline_alpha = 1,
  cust_theme = theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "regulation estimate",
  y_lab = "error",
  ...
)
}
\arguments{
\item{x}{an \code{\link{actiData}} object.}

\item{reactions}{a character vector with BiGG IDs of metabolic reactions
to be presented in the plot. If \code{NULL}, all reactions will be displayed.}

\item{subsystems}{a character vector with names of metabolic subsystems to
be presented in the plot. If \code{NULL}, all subsystems will be displayed.}

\item{fun}{a function used to transform the regulation estimates ad their
errors prior to plotting. Defaults to \code{identity}.}

\item{point_color}{fill color of data points.}

\item{point_size}{size of data points.}

\item{point_alpha}{alpha/opacity of data points.}

\item{show_trend}{logical, should a trend line be plotted
(with \code{\link[ggplot2]{geom_smooth}}).}

\item{line_color}{color of the trend line.}

\item{line_alpha}{alpha/opacity of the trend line.}

\item{show_vline}{logical: should a vertical line representing non-regulated
reactions (\code{fold_reg = 1}) be displayed in the plot?}

\item{vline_color}{color of the line representing no regulation.}

\item{vline_type}{type of the line representing no regulation.}

\item{vline_alpha}{alpha/opacity of the line representing no regulation.}

\item{show_hline}{logical: should a vertical line representing zero errors
(\code{error = 0}) be displayed in the plot?}

\item{hline_color}{color of the line representing zero errors.}

\item{hline_type}{type of the line representing zero errors.}

\item{hline_alpha}{alpha/opacity of the line representing zero errors.}

\item{cust_theme}{\code{NULL} or a custom \code{ggplot} theme.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{x_lab}{X axis label.}

\item{y_lab}{Y axis label.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_smooth}}.}
}
\value{
a \code{ggplot} scatter plot.
}
\description{
Function \code{plot_errors()} takes an \code{\link{actiData}} object with estimates
of regulation of activity of metabolic reactions and visualizes the reaction
activity estimates and their errors in a scatter plot.
}
\details{
The function works only for \code{\link{actiData}} objects with errors of
the regulation estimates obtained from the normal distribution or by Monte
Carlo simulations, i.e. objects generated with \code{\link{get_regulation}}
with the non-\code{NULL} \code{err} argument.
}
\seealso{
\code{\link[=get_regulation]{get_regulation()}}
}
