% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{plot_mc}
\alias{plot_mc}
\title{Plot reaction regulation estimates in iterations of Monte-Carlo simulations.}
\usage{
plot_mc(
  x,
  reactions = NULL,
  subsystems = NULL,
  fun = identity,
  type = c("line", "violin"),
  point_color = "steelblue",
  point_size = 2,
  point_alpha = 0.75,
  point_wjitter = 0,
  point_hjitter = 0.1,
  line_alpha = 1,
  line_width = 0.25,
  show_zero_line = TRUE,
  zero_line_color = "black",
  zero_line_type = "dashed",
  zero_line_alpha = 1,
  cust_theme = theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = NULL,
  ...
)
}
\arguments{
\item{x}{an \code{\link{actiData}} object.}

\item{reactions}{a character vector with BiGG IDs of metabolic reactions
to be presented in the plot. If \code{NULL}, all reactions will be displayed.}

\item{subsystems}{a character vector with names of metabolic subsystems to
be presented in the plot. If \code{NULL}, all subsystems will be displayed.}

\item{fun}{a function used to transform the regulation estimates ad their
errors prior to plotting. Defaults to \code{identity}.}

\item{type}{type of the plot, see Details.}

\item{point_color}{fill color of data points.}

\item{point_size}{size of data points.}

\item{point_alpha}{alpha/opacity of data points.}

\item{point_wjitter}{width of jittering of data points.}

\item{point_hjitter}{height of jittering of data points.}

\item{line_alpha}{alpha/opacity of lines representing single reactions
when \code{type = "line"}.}

\item{line_width}{width of lines representing single reactions when \code{type = "line"}.}

\item{show_zero_line}{logical should lines representing no regulation be
displayed in the plots (\code{fold_reg = 1})?}

\item{zero_line_color}{color of the lines of no regulation.}

\item{zero_line_type}{type of the no regulation line.}

\item{zero_line_alpha}{alpha/opacity of the no regulation lines.}

\item{cust_theme}{\code{NULL} or a custom \code{ggplot} theme.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{x_lab}{X axis label.}

\item{y_lab}{Y axis label.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_line}} and
\code{\link[ggplot2]{geom_violin}}.}
}
\value{
a \code{ggplot} graphic.
}
\description{
Function \code{plot_mc()} creates line plots of reaction regulation estimates in
iterations of the simulation algorithm, i.e. subsequent draws from normal
distribution of estimates of differential gene expression.
}
\details{
The function works only for \code{\link{actiData}} objects with Monte Carlo
estimates (i.e. created with \code{\link{get_regulation}} with
non \code{NULL} \code{err} vector, \verb{err_method = "mc}, and \code{return_mc = TRUE}).

Two graph forms are available:
\itemize{
\item line/path plots (\code{type = "line"}), where subsequent iterations of are presented
in the X axis and regulation estimates are presented in the Y axis,
estimates of the same reaction are connected with lines.
\item violin plots (\code{type = "violin"}), where reactions are presented in the
Y axis and estimates are presented in the X axis. Distribution of the
estimates is presented in violins, estimates in single reactions are shown
as points.
}
}
