% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{plot_numbers}
\alias{plot_numbers}
\title{Bar/stack plots with numbers of significantly activated and inhibited reactions.}
\usage{
plot_numbers(
  x,
  reactions = NULL,
  subsystems = NULL,
  scale = c("count", "percent"),
  type = c("plus_minus", "bar", "stack"),
  palette = c(activated = "firebrick", inhibited = "steelblue"),
  bar_rim_color = "black",
  show_all = FALSE,
  labeller_fun = identity,
  show_n_total = TRUE,
  cust_theme = microViz::theme_micro(),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = NULL,
  y_lab = "subsystem",
  fill_lab = "regulation\\nstatus",
  ...
)
}
\arguments{
\item{x}{an \code{\link{actiData}} object.}

\item{reactions}{a character vector with BiGG IDs of metabolic reactions
to be presented in the plot. If \code{NULL}, all reactions will be displayed.}

\item{subsystems}{a character vector with names of metabolic subsystems to
be presented in the plot. If \code{NULL}, all subsystems will be displayed.}

\item{scale}{type of the frequency statistic to be shown in the plot:
count (default) or percentage of reactions in the subsystem.}

\item{type}{type of the plot, see Details.}

\item{palette}{color palette: a named character vector with at least two elements
specifying the colors for activated and inhibited reactions.}

\item{bar_rim_color}{color of the bar's rim/line.}

\item{show_all}{logical, should frequencies of all regulated reactions be
presented in the plot together with counts/percentages for the subsystems?}

\item{labeller_fun}{a function used to transform subsystem names, e.g. into
abbreviations, which are shown in the Y axis of the plot.}

\item{show_n_total}{logical: if \code{TRUE}, total numbers of reactions in subsystems
are displayed in the Y axis along with the subsystem names.}

\item{cust_theme}{\code{NULL} or a custom \code{ggplot} theme.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{x_lab}{X axis label.}

\item{y_lab}{Y axis label.}

\item{fill_lab}{title of the fill scale.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_bar}}.}
}
\value{
a \code{ggplot} graphic.
}
\description{
Representation of numbers or percentages of differentially regulated
reactions in subsystems as bar or stack plot.
}
\details{
Please make sure that the object contains activity regulation status
information.
If not, please call \code{\link{identify_regulated}} prior to
counting the significant effects.

The plot type is determined by \code{type} argument:
\code{"plus_minus"} (bar plot, inhibited reaction as "negative" numbers),
\code{"bar"} (bar plot, dodged bars for inhibited and activated reactions),
\code{"stack"} plot.
}
