% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiData_methods.R
\name{print.actiData}
\alias{print.actiData}
\title{Print method for \code{actiData} objects.}
\usage{
\method{print}{actiData}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{actiData}}.}

\item{...}{additional arguments, currently none.}
}
\value{
nothing, called for its side effects.
}
\description{
Print method for \code{actiData} objects.
}
