% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{print.geneSBML}
\alias{print.geneSBML}
\title{Appearance for the geneSBML class.}
\usage{
\method{print}{geneSBML}(x, ...)
}
\arguments{
\item{x}{an object of the geneSBML class.}

\item{...}{extra arguments, currently none.}
}
\description{
Print method for the geneSBML class.
}
