% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_funs.R
\name{react_to_gene}
\alias{react_to_gene}
\title{Map reactions to gene EntrezID.}
\usage{
react_to_gene(react_id, geneSBML)
}
\arguments{
\item{react_id}{BiGG reaction ID, with or without the leading 'R_' string.}

\item{geneSBML}{a geneSBML object.}
}
\value{
a data frame with reaction ids and a list of the associated
gene Entrez IDs.
}
\description{
Maps reaction BiGG IDs to the Entrez IDs of the associated
genes.
}
