% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_funs.R
\name{react_to_metab}
\alias{react_to_metab}
\title{Map reactions to metabolites.}
\usage{
react_to_metab(
  react_id,
  annotation_db = biggrExtra::reactions,
  lead_str = TRUE,
  exc_regex = NULL,
  detailed = FALSE
)
}
\arguments{
\item{react_id}{BiGG reaction ID, with or without the leading 'R_' string.}

\item{annotation_db}{an annotation data frame with the 'bigg_id' and
'reaction_string' variables.}

\item{lead_str}{logical, should the metabolite ID contain the leading 'M_'
string? Defaults to TRUE.}

\item{exc_regex}{a regular expression to exclude
some of metabolites (such as water) from the output. Ignored if NULL.}

\item{detailed}{logical, should the output contain information on substrates
and products? Dafaluts to FALSE.}
}
\value{
a data frame with reaction IDs and a list
of the associated metabolite IDs. If 'detailed' is set to TRUE,
the colums 'lhs' and 'rhs' contain the left and right hand side metabolites
of the reaction.
}
\description{
Maps reaction BiGG IDs to BiGG IDs of
the associated metabolites.
}
