% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{reactions}
\alias{reactions}
\title{BiGG reaction annotation data.}
\format{
A data frame with 28301 rows and 6 variables:
\itemize{
\item \strong{bigg_id}: BiGG reaction identifier
\item \strong{name}: reaction name
\item \strong{reaction_string}: reaction equation
\item \strong{model_list}: list of BiGG models containing the reaction
\item \strong{database_links}: links to the BiGG on-line database
\item \strong{old_bigg_ids}: legacy BiGG ID
}
}
\source{
BiGG: http://bigg.ucsd.edu/data_access
}
\usage{
data(reactions)
}
\description{
A data frame with BiGG database annotation of metabolic
reactions.
}
