% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{Recon2_2D}
\alias{Recon2_2D}
\title{Reaction association rules for the Recon2.2 model.}
\format{
a data frame with 7785 rows and 7 columns:
\itemize{
\item \strong{id} and \strong{metaid} reaction identifiers
\item \strong{sboTerm} Systems Biology Ontology term
\item \strong{name} reaction name
\item \strong{gene_association} gene - reaction association rules as character strings
\item \strong{subsystem} Recon metabolic subsystem
\item \strong{confidence_level} confidence level for the gene - reaction association rule
}
}
\source{
\href{https://www.ebi.ac.uk/biomodels/MODEL1603150001}{biomodels} and
\href{https://github.com/PiotrTymoszuk/RECON_processing}{RECON_processing GitHub respository} .
}
\usage{
data(Recon2_2D)
}
\description{
Cleared and harmonized reaction annotation data for
\href{https://www.ebi.ac.uk/biomodels/MODEL1603150001}{Recon2.2 model of human metabolism}.
The data were processed with the \href{https://github.com/PiotrTymoszuk/RECON_processing}{RECON_processing}
R pipeline.
}
\references{
Swainston N, Smallbone K, Hefzi H, Dobson PD, Brewer J, Hanscho M, Zielinski DC,
Ang KS, Gardiner NJ, Gutierrez JM, et al.
Recon 2.2: from reconstruction to model of human metabolism.
Metabolomics (2016) 12. doi:10.1007/S11306-016-1051-4
}
