% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{Recon3D}
\alias{Recon3D}
\title{Corrected Recon3 SBML model.}
\format{
SBML model
}
\source{
BiGGR package, BiGGR database
(http://bigg.ucsd.edu/models/RECON1/reactions/ATPS4m).
}
\usage{
data(Recon3D)
}
\description{
A version of the SBML Recon3 model provided
by the BiGGR package with manual correction of errors in the gene
association rules (detected for the reaction R_ATPS4m, slot: notes,
missing parentheses, replacement with the rule from Recon1 model).
}
