% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_utils.R
\name{sbml_to_hd}
\alias{sbml_to_hd}
\title{Generate a hypergraph of the SMBL model.}
\usage{
sbml_to_hd(
  model,
  rates,
  rate_sep = ":",
  suffixes = NULL,
  suffix_sep = " ",
  colors = NULL,
  node_color = "gray30",
  signif_digits = 2,
  fontsize = 0.75,
  relevant.species = names(model@species),
  relevant.reactions = names(model@reactions),
  layoutType = "dot",
  lwd.max = 3,
  lwd.min = 0.5,
  plt.margins = c(150, 150, 150, 150)
)
}
\arguments{
\item{model}{a SBML model.}

\item{rates}{a named numeric vector with reaction regulation estimates
(linear, not log2). The names should be reaction IDs compatible with the
SBML model.}

\item{rate_sep}{a character separating the reaction identifier
and its regulation rate in the graph.}

\item{suffixes}{additional information to be displayed in the graph along
with the reactions and their rates. Omitted if NULL.}

\item{suffix_sep}{a character separating the reaction identifier/rate
from the suffix in the in the graph.}

\item{colors}{a vector of colors for the reactions. Should be named with
reaction IDs.}

\item{node_color}{default node color.}

\item{signif_digits}{significant digits used for rounding of the
reaction rate.}

\item{fontsize}{size of the font, in cex-units.}

\item{relevant.species}{metabolites to be plotted. If not specified by
the user, all metabolites participating in the reactions specified by
'relevant.reactions' are displayed.}

\item{relevant.reactions}{BiGG ID of the reactions to be plotted.}

\item{layoutType}{is a character string representing the layout engine to
be used for visualization. Current supported layouts are "dot", "twopi",
"neato","fdp","sfdp" and "circo". Defaults to "dot".}

\item{lwd.max}{a numeric given the maximum edge width. Defaults to 3.}

\item{lwd.min}{a numeric given the minimum edge width. Defaults to 0.5}

\item{plt.margins}{A numerical vector of the form
c(bottom, left, top, right) giving additional white space around the graph
(in case long node or edge labels fall outside the plotting region).
Defaults to c(150,150,150,150).}
}
\value{
a hypergraph object of the 'RagraphBPH' class. Plotted
conveniently with the plot() method.
}
\description{
Produces a hypergraph (an instance of 'RagraphBPH' class).
The code was accommodated from \code{\link[BiGGR]{sbml2hyperdraw}} with
few modifications allowing for labeling of confidence intervals,
errors and significance.
}
