% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiData_methods.R
\name{select.actiData}
\alias{select.actiData}
\title{Select reactions or subsystems.}
\usage{
\method{select}{actiData}(.data, reactions = NULL, subsystems = NULL, ...)
}
\arguments{
\item{reactions}{a character vector with BiGG IDs of metabolic reactions.}

\item{subsystems}{a character vector with names of metabolic subsystems.}

\item{...}{additional arguments passed to methods.}

\item{x}{a \code{\link{actiData}} object.}
}
\value{
an \code{\link{actiData}} object with reactions and/or subsystems
of interest.
}
\description{
Select reactions or subsystems in an \code{\link{actiData}}.
}
\seealso{
\code{\link[=get_regulation]{get_regulation()}}
}
