% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_utils.R
\name{sub_fisher}
\alias{sub_fisher}
\title{Test for enrichment in subsystems with Fisher's exact test.}
\usage{
sub_fisher(data, alternative = "greater", .parallel = FALSE, ...)
}
\arguments{
\item{.parallel}{logical, should the computation be run in parallel?
Defaults to FALSE.}
}
\value{
a data frame with subsystem name and testing results
for regulated, activated and inhibited reactions
as well as FDR-adjusted p values.
}
\description{
Tests for significant enrichment of generally significantly
regulated, significantly activated and significantly inhibited reactions in
metabolic subsystems.
The significance is determined by Fisher's exact test
(H0: no significantly regulated reactions)
}
\details{
The method is biased towards subsystems with multiple reactions!
Technically, Fisher's exact tests are computed with
\code{\link[rstatix]{fisher_test}}.
}
