% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_utils.R
\name{sub_simulation}
\alias{sub_simulation}
\title{Test for enrichment against random reaction draws.}
\usage{
sub_simulation(
  model,
  signif_type = c("fdr", "raw"),
  regulation_level = 1,
  n_iter = 1000,
  .parallel = FALSE
)
}
\arguments{
\item{model}{a geneSBML model.}

\item{signif_type}{type of significance to be used for definition of
regulated reactions:
'raw' (default) or 'fdr' (FDR-corrected).
Ignored if the model contains no error information.}

\item{regulation_level}{fold-regulation level cutoff used for definition
of regulated reactions.}

\item{n_iter}{number of random draws from the reaction set.}

\item{.parallel}{logical, should the computation be run in parallel?}

\item{aggregate}{logical, should the results be aggregated
to a single data frame?}
}
\description{
Tests for enrichment of significantly regulated reactions
within metabolic subsystems by comparing the actual counts with
n random draws from the reaction set.
}
