% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funs.R
\name{suba}
\alias{suba}
\title{Subsystem enrichment analysis for differentially regulated reactions.}
\usage{
suba(x, type = c("fisher", "random"), ...)
}
\arguments{
\item{x}{a \code{\link{actiData}} object.}

\item{type}{type of the hypothesis test: Fisher's exact test or random drawing
from the whole reaction pool.}

\item{...}{additional arguments passed to \code{\link[fastTest]{f_enrichment}}
which specify e.g. type of confidence intervals for the OR statistic and
number of random draws, and serial/parallel execution of the testing algorithm.}
}
\value{
a data frame with enrichment analysis results:
numbers of reactions in the activated and inhibited reaction sets and in
the subsystems, enrichment magnitude measured by odds ratio (OR) statistic,
raw p values, and p values adjusted for multiple testing.
}
\description{
Function \code{suba()} performs enrichment analyses for member reactions of
metabolic subsystems among significantly activated and inhibited reactions in
an \code{\link{actiData}} object.
Please note, that prior to the analyses, significantly activated and
inhibited reactions need to by determined by calling
\code{\link{identify_regulated}}.
}
\details{
The analyses are done with \code{\link[fastTest]{f_enrichment}} with two methods:
Fisher's exact test (\verb{type = "fisher}) or by random drawing from the entire
reaction pool (\code{type = "random"}).
}
\seealso{
\code{\link[=get_regulation]{get_regulation()}}
}
