% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactDB_methods.R
\name{summary.reactDB}
\alias{summary.reactDB}
\title{Summary method for \code{reactDB} objects.}
\usage{
\method{summary}{reactDB}(object, ...)
}
\arguments{
\item{object}{an \code{\link{reactDB}} object.}

\item{...}{additional arguments, currently none.}
}
\value{
A data frame with numbers of reactions, subsystems, genes, and gene - reaction
association rules.
}
\description{
\code{summary()} method for \code{reactDB} objects.
}
