% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{tcga_data}
\alias{tcga_data}
\title{Gene regulation estimates between collagen-subtypes of prostate carcinoma.}
\format{
A data frame with 36118 rows and 8 variables.
\itemize{
\item \strong{gene_symbol}: HGNC gene symbol
\item \strong{entrez_id}: Entrez ID identifier
\item \emph{estimate}: log~2~ regulation estimate versus Collagen^low^ subtype
\item \emph{se}: standard error of the mean for the regulation estimate
\item \strong{lower_ci}: lower 95\%CI interval limit for the regulation estimate
\item \strong{upper_ci}: upper 95\%CI interval limit for the regulation estimate
\item \strong{p_adjusted}: FDR-corrected p value.
\item \strong{regulation}: regulation sign.
}
}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/38851995/}{Heidegger et al. 2025}
}
\usage{
data(tcga_data)
}
\description{
A data frame with gene expression regulation estimates between
Collagen^hi^ and Collagen^low^ tumors. The estimates are log~2~
fold regulation estimates versus the Collagen^low^ tumor subtype computed by
linear regression. The values were calculated for the prostate cancer TCGA
(The Cancer Gene Atlas) data set.
}
\references{
Heidegger I, Frantzi M, Salcher S, Tymoszuk P, Martowicz A, Gomez-Gomez E,
Blanca A, Lendinez Cano G, Latosinska A, Mischak H, et al.
Prediction of Clinically Significant Prostate Cancer by a Specific Collagen-related
Transcriptome, Proteome, and Urinome Signature.
Eur Urol Oncol (2024) doi:10.1016/J.EUO.2024.05.014
}
