% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_mc}
\alias{test_mc}
\title{Regulation significance based on MC simulation results.}
\usage{
test_mc(fold_reg, mu = 1)
}
\arguments{
\item{fold_reg}{MC-simulated fold-regulation values.}

\item{mu}{the reference value.}
}
\value{
a numeric p value.
}
\description{
Tests for gene regulation significance based on
Monte Carlo (MC) simulation results.
}
