% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actiData_methods.R
\name{transform_estimates}
\alias{transform_estimates}
\alias{transform_estimates.actiData}
\title{Transform of regulation activity estimates.}
\usage{
transform_estimates(x, ...)

\method{transform_estimates}{actiData}(x, fun = log2, prefix = "log2_", ...)
}
\arguments{
\item{x}{a \code{\link{actiData}} object.}

\item{...}{additional arguments passed to methods.}

\item{fun}{transformation function.}

\item{prefix}{a character string preceding names of columns storing the
transformed estimates, errors, and confidence interval's bounds.}
}
\value{
a \code{\link{actiData}} object.
The \code{reg} element of the object is appended with additional with transformed
fold-regulation estimates, their errors, and bounds of the confidence intervals.
Names of these new column are preceded with a character string specified by
\code{prefix}
}
\description{
Transformation of estimates of regulation of reaction activity with a
function.
}
\details{
Function \code{transform_estimates} transforms estimates of reaction activity
along with their errors and confidence intervals (if present) with an
user-provided function (\code{log2} by default).
Note, no checks of validity of the function output are conducted!
}
\seealso{
\code{\link[=get_regulation]{get_regulation()}}
}
