% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{visualize.geneSBML}
\alias{visualize.geneSBML}
\title{Create a hypergraph from a geneSBML object.}
\usage{
\method{visualize}{geneSBML}(
  x,
  rate_sep = ":",
  suffixes = c("none", "error", "p_raw", "p_fdr"),
  suffix_sep = " ",
  colors = c(activated = "firebrick", inhibited = "steelblue", ns = "gray70"),
  node_color = "gray30",
  signif_type = c("raw", "fdr"),
  signif_digits = 2,
  fontsize = 0.75,
  relevant.species = components(x, "metabolites"),
  relevant.reactions = components(x, "reactions"),
  layoutType = "dot",
  lwd.max = 3,
  lwd.min = 0.5,
  plt.margins = c(150, 150, 150, 150),
  ...
)
}
\arguments{
\item{x}{a geneSBML object.}

\item{rate_sep}{a character separating the reaction identifier
and its regulation rate in the graph.}

\item{suffixes}{additional information to be displayed in the graph along
with the reactions and their rates: none ('none', default),
error ('error'), raw p value ('p_raw')
or FDR-adjusted p value ('p_fdr').}

\item{suffix_sep}{a character separating the reaction identifier/rate
from the suffix in the in the graph.}

\item{colors}{a named vector of colors for the reactions
(names: 'activated', 'inhibited', 'ns').}

\item{node_color}{default node color.}

\item{signif_type}{significance to be plotted in Volcano or Forest plots:
'raw' (default) or 'fdr' (FDR-corrected).}

\item{signif_digits}{significant digits used for rounding of numeric values
presented in the plot.}

\item{fontsize}{size of the font, in cex-units.}

\item{relevant.species}{metabolites to be plotted. If not specified by
the user, all metabolites participating in the reactions specified by
'relevant.reactions' are displayed.}

\item{relevant.reactions}{BiGG ID of the reactions to be plotted.
With or without leading 'R_' string.}

\item{layoutType}{is a character string representing the layout engine to
be used for visualization. Current supported layouts are "dot", "twopi",
"neato","fdp","sfdp" and "circo". Defaults to "dot".}

\item{lwd.max}{a numeric given the maximum edge width. Defaults to 3.}

\item{lwd.min}{a numeric given the minimum edge width. Defaults to 0.5}

\item{plt.margins}{A numerical vector of the form
c(bottom, left, top, right) giving additional white space around the graph
(in case long node or edge labels fall outside the plotting region).
Defaults to c(150,150,150,150).}

\item{...}{extra arguments, currently none.}
}
\description{
Generates a hypergraph (instance of the 'RagraphBPH' class)
for the requested reaction and metabolites.
}
\details{
Technically, a wrapper around \code{\link[BiGGR]{sbml2hyperdraw}}.
Rates for particular reactions and their errors are retrieved from the
regulation table stored within the object.
Reactions and their rates are presented in the edges. In addition, those
reaction labels may be appended by user-specified suffixes such as
confidence intervals or p values.
Edge width corresponds to absolute log~2~ fold regulation. The edge color is
specified by raw or FDR-corrected significance and the regulation sign.
}
