% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_funs.R
\name{annotate_bigg}
\alias{annotate_bigg}
\title{Retrieve metabolite or reaction features from a local BiGG database.}
\usage{
annotate_bigg(
  bigg_id,
  value = "name",
  annotation_db = reactions,
  id_type = c("reaction", "metabolite")
)
}
\arguments{
\item{bigg_id}{BiGG ID identifier of reactions or metabolites.
A character vector, with or without trailing 'R_' or 'M_' strings.}

\item{value}{the name of the annotation table output variable.}

\item{annotation_db}{annotation data frame or a SBML database.}

\item{id_type}{type of the identifier: reaction (default) or metabolite.
Ignored if annotation_db is a data frame.}
}
\value{
a vector with the requested features.
}
\description{
Retrieves the requested metabolite or reaction name from
a local database given the BiGG ID as a key.
}
\details{
You may use the 'metabolites' or 'reactions' data sets provided
by the package or download the latest annotation data sets
from http://bigg.ucsd.edu/data_access.
}
