% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_funs.R
\name{build_geneSBML}
\alias{build_geneSBML}
\title{Create a SBML model from gene Entrez IDs.}
\usage{
build_geneSBML(
  x,
  err = NULL,
  database = biggrExtra::Recon2D,
  scale = c("identity", "log2", "log"),
  or_fun = c("mean", "median", "min", "max"),
  and_fun = c("min", "max", "mean", "median"),
  x_default = 1,
  err_method = c("norm", "mc"),
  n_iter = 1000,
  mc_estimate = c("mean", "median"),
  burn_in = 0,
  ci_method = c("perc", "bca"),
  seed = NULL,
  save_memory = FALSE,
  .parallel = TRUE
)
}
\arguments{
\item{x}{a numeric vector with fold-regulation estimates. Elements are named
with Entrez IDs.}

\item{err}{a numeric vector with an error estimate such as SD or SEM, named
with Entrez IDs.}

\item{database}{an object of class \code{SBMLDocument} providing the gene -
reaction association rules.}

\item{scale}{specifies the scale of the provided expression regulation and
error estimates.}

\item{or_fun}{one of 'mean' (default), 'median', 'min' or 'max'. Specifies
the name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{and_fun}{one of 'mean', 'median', 'min' (default) or 'max'. Specifies
the name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{x_default}{the default numeric value or \code{NA} for regulation of the genes
present in the database and absent from the regulation data. If \code{NA},
the genes missing from x won't be included in subsequent SBML model.}

\item{err_method}{specifies the method of calculation of errors for reactions
as described in Details: 'norm' (from normal distribution, default) or
'mc' (Monte Carlo simulation).}

\item{n_iter}{number of iterations for the Monte Carlo simulation of
the reaction errors.}

\item{mc_estimate}{statistic used to compute the reaction activity estimate
in Monte Carlo simulation: either mean or median over algorithm iterations.
Ignored if no errors were provided or \code{err_method = 'norm'}.}

\item{burn_in}{number of Monte Carlo iterations to be discarded prior to
computation of activity estimates and activity regulation errors. Ignored
if no errors were provided or \code{err_method = 'norm'}.}

\item{ci_method}{method of calculation of confidence intervals:
percentile ('perc', default) or BCA ('bca').}

\item{seed}{a seed for the random number generator.}

\item{save_memory}{logical, should an abridged object without results of
single iterations of the Monte Carlo algorithm and without the SBML model
be returned?}

\item{.parallel}{logical, should th computation be run in parallel?}
}
\value{
a geneSBML object containing the following:
\itemize{
\item \code{model}: a rsbml Model built with the reaction IDs associated with the
provided genes, as described for
\code{\link[BiGGR]{buildSBMLFromReactionIDs}}.
\item \code{reg}: a data frame with reaction regulation estimates and their errors as
specified for \code{\link{get_regulation}}.
\item \code{gene_map}: a data frame with mapping of the reaction to gene identifiers
and association rules
\item \code{mc}: for models with regulation estimates and errors computed with the
Monte Carlo method, this component stores a matrix with results of
the simulation, i.e. activity regulation estimates for particular reactions
at algorithm iterations. Omitted when \code{save_memory = TRUE}.
}
}
\description{
Computes regulation estimates for the reactions present
in the database based on the regulation estimates for gene expression.
As described in the seminal BiGGR package paper
(DOI 10.1371/journal.pone.0119016), it is assumed that the change
in gene expression corresponds directly to the magnitude of modulation of
the pathway, e.g. two-fold increase in expression of an enzyme leads to a
two-fold increase in activity of the respective pathway
(gene - protein - reaction or GPR principle). In case, more genes
are mapped to a given pathway, the fold regulation is computed accordingly
to the logical operators between the genes.
If a given reaction is not associated with any genes, its regulation estimate
is set as specified by the \code{x_default} argument.
If a vector with regulation estimate errors is provided, the cumulative error
will be returned as well - see Details.
}
\details{
Works in a comparable fashion to \code{\link[BiGGR]{gprMapping}} but
considerably faster.
When a vector of errors for gene regulation is provided, there are two ways
to compute estimates and errors for the pathway regulation:
\itemize{
\item \strong{from the normal distribution}: the regulation estimates are computed
by evaluation of the gene - reaction association rules directly with
expression regulation estimates provided in \code{x}. The cumulative reaction
error is calculated as a sum of errors for the genes mapped to the reaction,
z statistic is computed as (fold-regulation - 1)/error,
raw p values and 95\% confidence intervals are obtained from the theoretical
distribution of z statistics and corrected for multiple testing with the
FDR/Benjamini-Hochberg method.
While this method is computationally fast, the error, t statistic,
confidence intervals and p values are just raw estimates of significance.
\item \strong{by Monte Carlo method}: given the gene expression regulation estimates
in \code{x} and their errors provided in \code{err}, a bunch of random values is drawn
from the normal distribution of with means specified by \code{x} and standard
deviations specified by \code{err}. Subsequently, activity regulation estimates
of the reactions are computed by evaluation of the gene - reaction
association rules for each set of simulated expression estimates. The
expected values of reaction activity regulation activity are derived as means
or medians across the algorithm iteration.
}
}
\references{
Gavai AK, Supandi F, Hettling H, Murrell P, Leunissen JAM, Van Beek JHGM.
Using Bioconductor Package BiGGR for Metabolic Flux Estimation Based on
Gene Expression Changes in Brain. PLoS One (2015) 10:e0119016.
doi:10.1371/JOURNAL.PONE.0119016

King ZA, Lu J, Dräger A, Miller P, Federowicz S, Lerman JA, Ebrahim A,
Palsson BO, Lewis NE. BiGG Models: A platform for integrating,
standardizing and sharing genome-scale models.
Nucleic Acids Res (2016) 44:D515–D522. doi:10.1093/NAR/GKV1049
}
