% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_funs.R
\name{build_geneSBML}
\alias{build_geneSBML}
\title{Create a SBML model from gene Entrez IDs.}
\usage{
build_geneSBML(
  x,
  err = NULL,
  database,
  or_fun = c("mean", "median", "min", "max"),
  and_fun = c("min", "max", "mean", "median"),
  x_default = 1,
  err_method = c("norm", "mc"),
  n_iter = 1000,
  ci_method = c("perc", "bca"),
  .parallel = TRUE
)
}
\arguments{
\item{x}{a numeric vector with fold-regulation estimates (not log2!!!).
Elements are named with Entrez IDs.}

\item{err}{a numeric vector with an error estimate such as SD or SEM
(not log2!!!), named with Entrez IDs.}

\item{database}{an object of class SBMLDocument.}

\item{or_fun}{one of 'mean' (default), 'median', 'min' or 'max'. Specifies the name of
the function used to handle the OR operator between the genes.}

\item{and_fun}{one of 'mean', 'median', 'min' (default) or 'max'. Specifies the name of
the function used to handle the OR operator between the genes.}

\item{x_default}{the default numeric value or NA for regulation of the genes
present in the database and absent from the regulation data. If NA,
the genes missing from x won't be included in subsequent SBML model creation.}

\item{err_method}{specifies the method of calculation of errors for reactions
as described in Details: 'norm' (from normal distribution, default) or
'mc' (Monte Carlo simulation).}

\item{n_iter}{number of iterations for the Monte carlo simulation of
the reaction errors.}

\item{ci_method}{method of calculation of confidence intervals:
percentile ('perc', default) or BCA ('bca').}
}
\value{
a geneSBML object containing the following:
\itemize{
\item a rsbml Model built with the reaction IDs associated with the provided genes,
as described for \code{\link[BiGGR]{buildSBMLFromReactionIDs}}.
\item a data frame with reaction regulation estimates and their errors as
specified for \code{\link{get_regulation}}.
}
}
\description{
Builds a SBML model given gene Entrez ID identifiers.
Technically, an up-stream function of \code{\link{get_regulation}}.
}
\details{
see: \code{\link{get_regulation}}.
}
