% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{components.geneSBML}
\alias{components.geneSBML}
\title{Extract geneSBML model components.}
\usage{
\method{components}{geneSBML}(
  object,
  type = c("model", "regulation", "gene_map", "reactions", "metabolites", "genes",
    "subsystems"),
  ...
)
}
\arguments{
\item{object}{a geneSBML object.}

\item{type}{type of the output, see above.}

\item{...}{extra arguments, currently none.}
}
\value{
the output specified by \code{type} argument.
}
\description{
Extracts the following geneSBML components:
SBML model (\code{type = 'model'}, default),
reaction regulation estimates ('regulation'),
gene mapping ('gene_map'),
all reaction IDs ('reactions'),
all metabolite IDs ('metabolites')
all gene IDs ('genes') or a data frame
with subsystem mapping ('subsystems').
}
