% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_funs.R
\name{extract_genes}
\alias{extract_genes}
\title{Extract gene identifiers from a database.}
\usage{
extract_genes(database, react_id = NULL)
}
\arguments{
\item{database}{an object of class SBMLDocument.}

\item{react_id}{BiGG reaction ID, with or without the leading 'R_' string.
Defaults to NULL, which means that all reactions are mapped to genes.}
}
\value{
a data frame containing reaction ID ('react_id'),
list of gene identifiers ('entrez_id') and a list of expressions to be
evaluated by at calculating reaction regulation estimates.
}
\description{
Extracts Entrez ID gene identifiers mapped to all or
user-specified reactions present in the given database SBML database.
}
\details{
A wrapper around \code{\link[BiGGR]{extractGeneAssociations}}.
Entrez ID identifier version is discarded silently. Reactions with
non-Entrez ID features mapped are removed from the output and
a parsing warning is raised.
}
