% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{geneSBML}
\alias{geneSBML}
\title{Create a geneSBML object.}
\usage{
geneSBML(model, reg, gene_map, mc = NULL)
}
\arguments{
\item{model}{a SBML model.}

\item{reg}{a data frame with reaction IDs ('react_id'), regulation estimates
('fold_reg') and, optionally, regulation errors ('error').}

\item{gene_map}{a data frame with reaction IDs ('react_id'),
list of Entrez ID identifiers ('entrez_id')
and evaluation expressions ('exprs').}

\item{mc}{optional, a matrix with reaction regulation estimates in each step
of Monte Carlo simulation. Columns are reactions, rows are subsequent runs.}
}
\value{
an instance of geneSBML class.
}
\description{
Creates a geneSBML object given a SBML model object and a
data frame with reaction regulation estimates and, optionally, regulation
errors, and a data frame with gene - reaction mapping.
}
