% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_regulation}
\alias{get_regulation}
\title{Calculate reaction regulation estimated based on gene expression.}
\usage{
get_regulation(
  x,
  err = NULL,
  database = biggrExtra::Recon2D,
  scale = c("identity", "log2", "log"),
  or_fun = c("mean", "median", "min", "max"),
  and_fun = c("min", "max", "mean", "median"),
  x_default = 1,
  return_genes = FALSE,
  return_mc = FALSE,
  err_method = c("norm", "mc"),
  n_iter = 1000,
  mc_estimate = c("mean", "median"),
  burn_in = 0,
  ci_method = c("perc", "bca"),
  seed = NULL,
  .parallel = TRUE
)
}
\arguments{
\item{x}{a numeric vector with fold-regulation estimates. Elements are named
with Entrez IDs.}

\item{err}{a numeric vector with an error estimate such as SD or SEM, named
with Entrez IDs.}

\item{database}{an object of class \code{SBMLDocument} providing the gene -
reaction association rules.}

\item{scale}{specifies the scale of the provided expression regulation and
error estimates.}

\item{or_fun}{one of 'mean' (default), 'median', 'min' or 'max'. Specifies
the name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{and_fun}{one of 'mean', 'median', 'min' (default) or 'max'. Specifies
the name of the function used to handle the \code{OR} operator between the genes in
the association rule.}

\item{x_default}{the default numeric value or \code{NA} for regulation of the genes
present in the database and absent from the regulation data. If \code{NA},
the genes missing from x won't be included in subsequent SBML model.}

\item{return_genes}{logical, should a data frame with assignment of the
genes to reactions be returned? Defaults to \code{FALSE}.}

\item{return_mc}{logical, should all Monte Carlo-determined regulation
estimates for reactions be returned? Defaults to \code{FALSE}.}

\item{err_method}{specifies the method of calculation of errors for reactions
as described in Details: 'norm' (from normal distribution, default) or
'mc' (Monte Carlo simulation).}

\item{n_iter}{number of iterations for the Monte Carlo simulation of
the reaction errors.}

\item{mc_estimate}{statistic used to compute the reaction activity estimate
in Monte Carlo simulation: either mean or median over algorithm iterations.
Ignored if no errors were provided or \code{err_method = 'norm'}.}

\item{burn_in}{number of Monte Carlo iterations to be discarded prior to
computation of activity estimates and activity regulation errors. Ignored
if no errors were provided or \code{err_method = 'norm'}.}

\item{ci_method}{method of calculation of confidence intervals:
percentile ('perc', default) or BCA ('bca').}

\item{seed}{a seed for the random number generator.}

\item{.parallel}{logical, should th computation be run in parallel?}
}
\value{
a data frame with the following variables: reaction identifiers
and the fold regulation estimates. If err is not NULL, an error estimate per
reaction is included in the data frame as well along with the z statistic
(fold-regulation/error), 95\% confidence intervals and
raw p-values (normal distribution) and FDR-corrected p-values.
If return_genes is set to TRUE, a gene-mapping table is returned as well in
a list with the fold regulation data frame.
}
\description{
Intended for internal use, see: \code{\link{build_geneSBML}}.
}
