% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_oop.R
\name{is_geneSBML}
\alias{is_geneSBML}
\title{Test for the geneSBML class instance.}
\usage{
is_geneSBML(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests whether the object is an instance of the geneSBML class.
}
