% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{Recon2D}
\alias{Recon2D}
\title{Reaction association rules for the Recon2 model.}
\format{
a data frame with 7440 rows and 10 columns:
\itemize{
\item \strong{id} and \strong{metaid} BiGG reaction identifiers
\item \strong{sboTerm} Systems Biology Ontology term
\item \strong{name} reaction name
\item \strong{gene_association} gene - reaction association rules as character strings
\item \strong{subsystem} Recon metabolic subsystem
\item \strong{ec_number} EC identifier of enzymatic reaction
\item \strong{confidence_level} confidence level for the gene - reaction association rule
\item \strong{reference} literature references
\item \strong{notes} additional comments
}
}
\source{
\href{https://www.ebi.ac.uk/biomodels/MODEL1109130000}{biomodels} and
\href{https://github.com/PiotrTymoszuk/RECON_processing}{RECON_processing GitHub repository} .
}
\usage{
data(Recon2D)
}
\description{
Cleared and harmonized reaction annotation data for
\href{https://www.ebi.ac.uk/biomodels/MODEL1109130000}{Recon2 model of human metabolism}.
The data were processed with the \href{https://github.com/PiotrTymoszuk/RECON_processing}{RECON_processing}
R pipeline.
}
\references{
Thiele I, Swainston N, Fleming RMT, Hoppe A, Sahoo S, Aurich MK,
Haraldsdottir H, Mo ML, Rolfsson O, Stobbe MD, et al.
A community-driven global reconstruction of human metabolism.
Nat Biotechnol 2013 315 (2013) 31:419–425. doi:10.1038/nbt.2488
}
