% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{Recon2D}
\alias{Recon2D}
\title{Corrected Recon2 SBML model.}
\format{
SBML model
}
\source{
BiGGR package, BiGGR database
(http://bigg.ucsd.edu/models/RECON1/reactions/ATPS4m).
}
\usage{
data(Recon2D)
}
\description{
A version of the SBML Recon2 model provided
by the BiGGR package with manual correction of errors in the gene association rules
(detested for the reaction R_ATPS4m, slot: notes, missing parentheses,
replacement with the rule from Recon1 model).
}
