% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funs.R
\name{suba}
\alias{suba}
\title{Test for enrichment of significantly regulated reactions
in metabolic subsystems.}
\usage{
suba(
  model,
  signif_type = c("raw", "fdr"),
  regulation_level = 1,
  method = c("fisher", "simulation"),
  alternative = "greater",
  n_iter = 1000,
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{model}{a geneSBML model.}

\item{signif_type}{type of significance to be used for definition of
regulated reactions:
'raw' (default) or 'fdr' (FDR-corrected).
Ignored if the model contains no error information.}

\item{regulation_level}{fold-regulation level cutoff used for definition
of regulated reactions.}

\item{method}{'fisher' or 'simulation' as described above.
Defaults to 'fisher'.}

\item{alternative}{indicates the alternative hypothesis
and must be one of "two.sided", "greater" or "less". Defaults to "greater".
Applies only to the 'fisher' method.}

\item{n_iter}{number of random draws from the reaction set.}

\item{.parallel}{logical, should the analysis be run in parallel?}

\item{...}{extra arguments passed to \code{\link[rstatix]{fisher_test}}.}
}
\value{
a data frame with subsystem name and testing results
for regulated, activated and inhibited reactions
as well as FDR-adjusted p values.
}
\description{
Tests for enrichment of significantly regulated reactions
within metabolic subsystems. This is done with one of two methods:
'fisher' using Fisher's exact test or 'simulation' where the observed
enrichment is compared with random draws from the reaction set.
}
\details{
Technically, Fisher's tests are done with
\code{\link[rstatix]{fisher_test}}. This method is biased
towards subsystems with multiple reactions mapped to them
(like fatty acid oxidation).
}
